Split out processHtml.ts
